usb: dwc3: Fix List Corruption seen with DCP detection
The re-initialization of the ext_chg_wait after DCP detection can
potentially corrupt the list used by the completion struct, especially
if there is a wait_for_completion triggered after DCP detection in the
dwc3_resume_work. It is unnecessary to re-init the completion when it is
already done in the dwc3_msm_probe. Hence just signal complete when
DCP is detected.
Propagation of (CR)

Change-Id: I0017be47f3f4a1acc18dc587dde1b912c879e0eb
Signed-off-by: Ravi Vembu <a18394@motorola.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>