mondrian: enable fstack-protector-strong
Enable protection against stack corruption. (most) Functions with the
possibility of stack corruption  are protected with a canary.

With this change, kernel size may increase by 2.3%. Performance loss is considered minimal. Security
against stack overflow is greatly improved. [1]

This is the improved version of CONFIG_CC_STACKPROTECTOR=y which is
enabled in shamu's kernel.

[1] https://lwn.net/Articles/584225

Bug: 28837708