usb: dwc3-tabpro: Sync with dwc3 dwc3-otg.configuration
Based on:

usb: dwc3_otg: Turn on VBUS before starting host

Currently the VBUS regulator is enabled only after initializing
the xHCI host stack, which can result in enumeration failures
for certain devices. This is likely due to VBUS having a
finite rise time when it is turned on, especially when a load
is already attached. Mitigate this by enabling VBUS first before
starting the host stack.

CRs-Fixed: 599498
Change-Id: I9785ae4682d0a2dfbcec921b869fc03550e0db82
Signed-off-by: Jack Pham <jackp@codeaurora.org>

https://github.com/LineageOS/android_kernel_samsung_klte/commit/18321fade6640da5c19433a67cb1e81dff61415e