usb: dwc3-tabpro: Increase EP0 bounce buffer size to 2KB
Mirror commit d09ba5f7ee0abcd9b89e44043247d00e1ad6301a to dwc3-tabpro directory

There is a known limitation in the DWC3 controller that EP0
control transfers must an even multiple of wMaxPacketSize (64 bytes
in high speed, 512 bytes in super speed). In the current handling
of an OUT data stage, if the length of the request buffer is not
aligned, a pre-reserved bounce buffer is used for that transfer
and copied back to the function when completed. The size of this
buffer is #defined as 512 bytes, and a WARN_ON is issued if the
request is greater than this since that indicates a buffer overrun
condition. A simple way to address this is to bump up this size
to 2048 to handle occassional large control transfers.

Change-Id: I4ea0ae0496bc39f14a384b50cf984b9456b0e3ec
Signed-off-by: Jack Pham <jackp@codeaurora.org>